Add backport for weakref.finalize to urllib3
This commit introduces a backport of the `weakref.finalize` method from Python 3 to the urllib3 package. The `weakref_finalize` class allows for the finalization of weak-referencable objects, enabling a specified function to be called when the object is garbage collected. This implementation includes support for atexit registration, ensuring that any remaining finalizers are executed upon program exit. The new functionality enhances resource management and cleanup capabilities within the urllib3 library.